+++ game_patch/multi/server.cpp
@@ -134,6 +134,10 @@ void load_additional_server_config(rf::Parser& parser)
         g_additional_server_config.anticheat_level = parser.parse_int();
     }
 
+    if (parser.parse_optional("$DF Semi Auto Minimum Fire Wait:")) {
+        g_additional_server_config.click_limiter_fire_wait = parser.parse_int();
+    }
+
     if (parser.parse_optional("$DF Require Client Mod:")) {
         g_additional_server_config.require_client_mod = parser.parse_bool();
     }