+++ docs/CHANGELOG.md
@@ -39,6 +39,8 @@ Version 1.9.0 (not released yet)
 - Add Kill Reward settings for dedicated servers
 - Do not load unnecessary VPPs in dedicated server mode
 - Add level filename to "Level Initializing" console message
+- Fix some fire packets being wrongfully dropped by click limiter (DF bug)
+- Add `$DF Semi Auto Click Rate Limit` option in dedicated server config
 - Properly handle WM_PAINT in dedicated server, may improve performance (DF bug)
 
 [@GooberRF](https://github.com/GooberRF)